added json_set_encoder
