+++ app.py
@@ -4,13 +4,9 @@ from scripts.message import Message
 from scripts.analysis import Analysis
 from io import StringIO
 import json
+from scripts.json_set_encoder import SetEncoder
 
 
-class SetEncoder(json.JSONEncoder):
-    def default(self, obj):
-        if isinstance(obj, set):
-            return list(obj)
-        return json.JSONEncoder.default(self, obj)
 
 st.title("WhatsApp Stats")
 st.markdown(''' #### _Discover the **secrets** behind your chats!_
@@ -44,7 +40,7 @@ else:
     for line in stringio:
         chatline = line.strip()
         if Message.is_valid_message(Message, line): # check if the line if valid message
-            message = Message(line) # problema in def_author
+            message = Message(line) 
             analysis.update_stats(message)
             last_message_analyzed = message
         else:
